refactor(ui): group mosaic organization profile files under a feature directory
