+++ packages/swingset/src/stories/organization-profile.stories.tsx
@@ -1,5 +1,5 @@
 /** @jsxImportSource @emotion/react */
-import { OrganizationProfileView } from '@clerk/ui/mosaic/aio/organization-profile-view';
+import { OrganizationProfileView } from '@clerk/ui/mosaic/organization/organization-profile-view';
 
 import type { StoryMeta } from '@/lib/types';
 
@@ -8,7 +8,7 @@ import { Default as OrganizationProfileGeneralPanelDemo } from './organization-p
 export const meta: StoryMeta = {
   group: 'Organization',
   title: 'OrganizationProfile',
-  source: 'packages/ui/src/mosaic/aio/organization-profile.tsx',
+  source: 'packages/ui/src/mosaic/organization/organization-profile.tsx',
 };
 
 export function Default() {
