+++ packages/ui/src/mosaic/organization/__tests__/organization-profile-delete-section.view.test.tsx
